refactor crate
- move api functions into ZebedeeClient impl
- use &self in api functions, to avoid unnecessary cloning of string & reqwest::Client in ZebedeeClient
- Added missing Doc comments to api functions
- moved tests & struct definitions in separate files
