--- src/ln_address.rs
@@ -1,324 +1,0 @@
-use crate::{StdResp, ZebedeeClient};
-use anyhow::Result;
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use validator::Validate;
-
-#[derive(Debug, Validate, Deserialize)]
-pub struct LnAddress {
-    #[validate(email)]
-    pub address: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnPayerData {
-    pub name: HashMap<String, bool>,
-    pub identifier: HashMap<String, bool>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnValidateMetadata {
-    #[serde(rename = "minSendable")]
-    pub min_sendable: u64,
-    #[serde(rename = "maxSendable")]
-    pub max_sendable: u64,
-    #[serde(rename = "commentAllowed")]
-    pub comment_allowed: u64,
-    pub tag: String,
-    pub metadata: String,
-    pub callback: String,
-    #[serde(rename = "payerData")]
-    pub payer_data: LnPayerData,
-    pub disposable: bool,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnValidateData {
-    pub valid: bool,
-    pub metadata: LnValidateMetadata,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnInvoice {
-    pub uri: String,
-    pub request: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnFetchChargeData {
-    #[serde(rename = "lnaddress")]
-    pub ln_address: String,
-    pub amount: String,
-    pub invoice: LnInvoice,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnSendPaymentData {
-    pub id: String,
-    pub fee: Option<String>,
-    pub unit: String,
-    pub amount: String,
-    pub preimage: Option<String>,
-    pub status: String,
-    pub invoice: String,
-    #[serde(rename = "walletId")]
-    pub wallet_id: String,
-    #[serde(rename = "transactionId")]
-    pub transaction_id: String,
-    #[serde(rename = "createdAt")]
-    pub created_at: DateTime<Utc>,
-    #[serde(rename = "processedAt")]
-    pub processed_at: DateTime<Utc>,
-    #[serde(rename = "callbackURL")]
-    pub callback_url: Option<String>,
-    #[serde(rename = "internalId")]
-    pub internal_id: Option<String>,
-}
-
-/// Use this struct to create a well crafted json body for your Lightning Address payments
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnPayment {
-    #[serde(rename = "lnAddress")]
-    pub ln_address: String,
-    pub amount: String,
-    pub comment: String,
-}
-
-impl Default for LnPayment {
-    fn default() -> Self {
-        LnPayment {
-            ln_address: String::from(""),
-            amount: String::from(""),
-            comment: String::from("using zebedee rust sdk"),
-        }
-    }
-}
-
-/// Use this struct to create a well crafted json body for creating charges for Ligthning Addresses
-#[derive(Debug, Serialize, Deserialize)]
-pub struct LnFetchCharge {
-    #[serde(rename = "lnaddress")]
-    pub ln_address: String,
-    pub amount: String,
-    pub description: String,
-}
-
-impl Default for LnFetchCharge {
-    fn default() -> Self {
-        LnFetchCharge {
-            ln_address: String::from(""),
-            amount: String::from(""),
-            description: String::from("using zebedee rust sdk"),
-        }
-    }
-}
-
-pub async fn pay_ln_address(
-    client: ZebedeeClient,
-    payment: LnPayment,
-) -> Result<StdResp<Option<LnSendPaymentData>>, anyhow::Error> {
-    let url = format!("{}/v0/ln-address/send-payment", client.domain);
-    let resp = client
-        .reqw_cli
-        .post(&url)
-        .header("Content-Type", "application/json")
-        .header("apikey", client.apikey)
-        .json(&payment)
-        .send()
-        .await?;
-
-    let status_code = resp.status();
-    let status_success = resp.status().is_success();
-    let resp_text = resp.text().await?;
-
-    if !status_success {
-        return Err(anyhow::anyhow!(
-            "Error: status {}, message: {}, url: {}",
-            status_code,
-            resp_text,
-            &url,
-        ));
-    }
-
-    let resp_serialized = serde_json::from_str(&resp_text);
-
-    let resp_seralized_2 = match resp_serialized {
-        Ok(c) => c,
-        Err(e) => {
-            return Err(anyhow::anyhow!(
-                "Was given a good status, but something failed when parsing to json\nserde parse error: {}, \ntext from API: {}\n status code: {}",
-                e,
-                resp_text,
-                status_code
-            ))
-        }
-    };
-
-    Ok(resp_seralized_2)
-}
-
-pub async fn fetch_charge_ln_address(
-    client: ZebedeeClient,
-    payment: LnFetchCharge,
-) -> Result<StdResp<Option<LnFetchChargeData>>, anyhow::Error> {
-    let url = format!("{}/v0/ln-address/fetch-charge", client.domain);
-    let resp = client
-        .reqw_cli
-        .post(&url)
-        .header("Content-Type", "application/json")
-        .header("apikey", client.apikey)
-        .json(&payment)
-        .send()
-        .await?;
-
-    let status_code = resp.status();
-    let status_success = resp.status().is_success();
-    let resp_text = resp.text().await?;
-
-    if !status_success {
-        return Err(anyhow::anyhow!(
-            "Error: status {}, message: {}, url: {}",
-            status_code,
-            resp_text,
-            &url,
-        ));
-    }
-
-    let resp_serialized = serde_json::from_str(&resp_text);
-
-    let resp_seralized_2 = match resp_serialized {
-        Ok(c) => c,
-        Err(e) => {
-            return Err(anyhow::anyhow!(
-                "Was given a good status, but something failed when parsing to json\nserde parse error: {}, \ntext from API: {}\n status code: {}",
-                e,
-                resp_text,
-                status_code
-            ))
-        }
-    };
-
-    Ok(resp_seralized_2)
-}
-
-pub async fn validate_ln_address(
-    client: ZebedeeClient,
-    lightning_address: LnAddress,
-) -> Result<StdResp<Option<LnValidateData>>, anyhow::Error> {
-    match lightning_address.validate() {
-        Ok(_) => (),
-        Err(e) => {
-            return Err(anyhow::anyhow!(
-                "Bad LN Address {}, ValidationError {}",
-                &lightning_address.address,
-                e,
-            ))
-        }
-    };
-
-    let url = format!(
-        "{}/v0/ln-address/validate/{}",
-        client.domain, lightning_address.address
-    );
-    let resp = client
-        .reqw_cli
-        .get(&url)
-        .header("Content-Type", "application/json")
-        .header("apikey", client.apikey)
-        .send()
-        .await?;
-
-    let status_code = resp.status();
-    let status_success = resp.status().is_success();
-    let resp_text = resp.text().await?;
-
-    if !status_success {
-        return Err(anyhow::anyhow!(
-            "Error: status {}, message: {}, url: {}",
-            status_code,
-            resp_text,
-            &url,
-        ));
-    }
-
-    let resp_serialized = serde_json::from_str(&resp_text);
-
-    let resp_seralized_2 = match resp_serialized {
-        Ok(c) => c,
-        Err(e) => {
-            return Err(anyhow::anyhow!(
-                "Was given a good status, but something failed when parsing to json\nserde parse error: {}, \ntext from API: {}\n status code: {}",
-                e,
-                resp_text,
-                status_code
-            ))
-        }
-    };
-
-    Ok(resp_seralized_2)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::env;
-
-    #[tokio::test]
-    async fn test_pay_ln_address() {
-        let apikey: String = env::var("ZBD_API_KEY").unwrap();
-        let zbdenv: String =
-            env::var("ZBD_ENV").unwrap_or_else(|_| String::from("https://api.zebedee.io"));
-        let zebedee_client = ZebedeeClient::new().domain(zbdenv).apikey(apikey).build();
-
-        let payment = LnPayment {
-            ln_address: String::from("miketwenty1@zbd.gg"),
-            amount: String::from("1000"),
-            ..Default::default()
-        };
-        let r = pay_ln_address(zebedee_client, payment)
-            .await
-            .unwrap()
-            .success;
-        assert!(r);
-    }
-    #[tokio::test]
-    async fn test_fetch_charge_ln_address() {
-        let apikey: String = env::var("ZBD_API_KEY").unwrap();
-        let zbdenv: String =
-            env::var("ZBD_ENV").unwrap_or_else(|_| String::from("https://api.zebedee.io"));
-        let zebedee_client = ZebedeeClient::new().domain(zbdenv).apikey(apikey).build();
-
-        let payment = LnFetchCharge {
-            ln_address: String::from("miketwenty1@zbd.gg"),
-            amount: String::from("1000"),
-            ..Default::default()
-        };
-        let r = fetch_charge_ln_address(zebedee_client, payment)
-            .await
-            .unwrap()
-            .success;
-        assert!(r);
-    }
-
-    #[tokio::test]
-    async fn test_validate_ln_address() {
-        let apikey: String = env::var("ZBD_API_KEY").unwrap();
-        let zbdenv: String =
-            env::var("ZBD_ENV").unwrap_or_else(|_| String::from("https://api.zebedee.io"));
-        let zebedee_client = ZebedeeClient::new().domain(zbdenv).apikey(apikey).build();
-
-        let ln_address = String::from("andre@zbd.gg");
-
-        let r = validate_ln_address(
-            zebedee_client,
-            LnAddress {
-                address: ln_address,
-            },
-        )
-        .await
-        .unwrap()
-        .success;
-        assert!(r);
-    }
-}
