+++ src/wallet/types.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use crate::StdResp;
+
+pub type WalletInfoResponse = StdResp<Option<WalletData>>;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct WalletData {
+    pub unit: String,
+    pub balance: String,
+}
