+++ src/withdrawal_request/mod.rs
@@ -1,0 +1,4 @@
+mod types;
+pub use types::*;
+#[cfg(test)]
+mod tests;
